api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
selectolax==0.3.21
SQLAlchemy==2.0.36
pydantic==2.9.2
python-dotenv==1.0.1
python-dateutil==2.9.0post0